新增會員刪除功能
1.修正會員刪除功能

2.改正ProductList刪除功能bug

+++ ShoppingCartPractice/Models/Repository/ProductListRepository.cs
@@ -49,8 +49,12 @@ namespace ShoppingCartPractice.Models.Repository
         }
         public void Delete(Carts carts)
         {
-            db.Carts.Remove(carts);
-            db.SaveChanges();
+            if (carts != default(Models.Carts))
+            {
+                db.Entry(carts).State = System.Data.Entity.EntityState.Deleted;
+                db.Carts.Remove(carts);
+                db.SaveChanges();
+            }
         }
     }
 }
